Single post view
Added @app.route('/post/<slug>') to process single posts. Helper
functions loop through all files and match the submitted slug,
which is built from the post title using slugify().

A boolean is returned if the title matches the slug and the text
file is processed and returned to the template as HTML.